+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.session.update.R
@@ -4,7 +4,8 @@ concerto.session.update = function(){
   sql = sprintf("UPDATE TestSession SET
     status = '%s',
     timeLimit = '%s',
-    error = '%s'
+    error = '%s',
+    updated = CURRENT_TIMESTAMP
     WHERE id='%s'",
   dbEscapeStrings(concerto$connection, toString(concerto$session$status)),
   dbEscapeStrings(concerto$connection, toString(concerto$session$timeLimit)),